chore: Ignore migrations and spreadsheets in .gitignore
The .gitignore file was updated to ignore the `api/migrations/` directory and the `spreadsheets/setor_economico - Copia.xlsx` and `spreadsheets/setor_economico.xlsx` files. This ensures that these files are not tracked by Git.

The `lookup_field` attribute was added to the `SubclasseViewSet` class in `api/views.py`. This specifies that the `cd_subclasse` field should be used as the lookup field for retrieving instances of the `Subclasse` model.

```